rewrite code
