+++ frontend/src/containers/CourseCreate/CourseCreate.tsx
@@ -2,9 +2,10 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import KakaoMap from '../../components/Map/KakaoMap';
+import poisData from '../../components/poisData.json';
 import SearchBar from '../../components/SearchBar/SearchBar';
 import { AppDispatch } from '../../store';
-import { courseSlice, fetchPathFromTMap, selectCourse } from '../../store/slices/course';
+import { fetchPathFromTMap, selectCourse } from '../../store/slices/course';
 import styles from './CourseCreate.module.scss';
 
 const { Tmapv3 } = window as any;
@@ -47,7 +48,8 @@ export interface MarkerProps {
 
 export default function CourseCreate() {
   const [map, setMap] = useState<kakao.maps.Map>();
-  const [markers, setMarkers] = useState<MarkerProps[]>([]);
+  const [searchMarkers, setSearchMarkers] = useState<MarkerProps[]>([]);
+  const [previewMarkers, setPreviewMarkers] = useState<MarkerProps[]>([]);
   const [path, setPath] = useState<PositionProps[]>([]);
   const [info, setInfo] = useState<MarkerProps | null>(null);
   const [selected, setSelected] = useState<MarkerProps[]>([]);
@@ -81,7 +83,7 @@ export default function CourseCreate() {
             });
             bounds.extend(new kakao.maps.LatLng(Number(data[i].y), Number(data[i].x)));
           }
-          setMarkers(markerArr);
+          setSearchMarkers(markerArr);
           setPreview(false);
 
           // 검색된 장소 위치를 기준으로 지도 범위를 재설정합니다
@@ -94,84 +96,88 @@ export default function CourseCreate() {
       });
     }
   };
+  const processData = () => {
+    const drawInfoArr = [];
+    const resultMarkerArr = [];
+    const bounds = new kakao.maps.LatLngBounds();
+    let viaPoint = 1;
 
-  // const processData = () => {
-  //   const drawInfoArr = [];
-  //   const resultMarkerArr = [];
-  //   const bounds = new kakao.maps.LatLngBounds();
-  //   let viaPoint = 1;
-  //
-  //   for (const i in resultFeatures) {
-  //     const { geometry } = resultFeatures[i];
-  //     const { properties } = resultFeatures[i];
-  //
-  //     if (geometry.type === 'LineString') {
-  //       for (const j in geometry.coordinates) {
-  //         // 경로들의 결과값(구간)들을 포인트 객체로 변환
-  //         const latlng = new Tmapv3.Point(geometry.coordinates[j][0], geometry.coordinates[j][1]);
-  //         // 포인트 객체를 받아 좌표값으로 변환
-  //         const convertPoint = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(latlng);
-  //         // console.log("convertPoint:", convertPoint);
-  //         // 포인트객체의 정보로 좌표값 변환 객체로 저장
-  //         const convertChange = { lat: convertPoint._lat, lng: convertPoint._lng };
-  //         // console.log("convertChange:", convertChange);
-  //
-  //         drawInfoArr.push(convertChange);
-  //       }
-  //     } else {
-  //       let markerImg = '';
-  //
-  //       if (properties.pointType === 'S') {
-  //         // 출발지 마커
-  //         markerImg = 'http://tmapapi.sktelecom.com/upload/tmap/marker/pin_r_m_s.png';
-  //       } else if (properties.pointType === 'E') {
-  //         // 도착지 마커
-  //         markerImg = 'http://tmapapi.sktelecom.com/upload/tmap/marker/pin_r_m_e.png';
-  //       } else {
-  //         // 각 포인트 마커
-  //         markerImg = `http://tmapapi.sktelecom.com/upload/tmap/marker/pin_b_m_${viaPoint}.png`;
-  //         viaPoint += 1;
-  //       }
-  //
-  //       // 경로들의 결과값들을 포인트 객체로 변환
-  //       const latlon = new Tmapv3.Point(geometry.coordinates[0], geometry.coordinates[1]);
-  //       // 포인트 객체를 받아 좌표값으로 다시 변환
-  //       const convertPoint = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(latlon);
-  //
-  //       resultMarkerArr.push({
-  //         position: {
-  //           lat: convertPoint._lat,
-  //           lng: convertPoint._lng,
-  //         },
-  //         content: properties.viaPointName,
-  //         image: markerImg,
-  //       });
-  //       bounds.extend(new kakao.maps.LatLng(convertPoint._lat, convertPoint._lng));
-  //     }
-  //   }
-  //   // 검색된 장소 위치를 기준으로 지도 범위를 재설정합니다
-  //   map?.setBounds(bounds);
-  //   setPath(drawInfoArr);
-  //   setMarkers(resultMarkerArr);
-  // };
+    for (const i in resultFeatures) {
+      const { geometry } = resultFeatures[i];
+      const { properties } = resultFeatures[i];
+
+      if (geometry.type === 'LineString') {
+        for (const j in geometry.coordinates) {
+          // 경로들의 결과값(구간)들을 포인트 객체로 변환
+          const latlng = new Tmapv3.Point(geometry.coordinates[j][0], geometry.coordinates[j][1]);
+          // 포인트 객체를 받아 좌표값으로 변환
+          const convertPoint = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(latlng);
+          // console.log("convertPoint:", convertPoint);
+          // 포인트객체의 정보로 좌표값 변환 객체로 저장
+          const convertChange = { lat: convertPoint._lat, lng: convertPoint._lng };
+          // console.log("convertChange:", convertChange);
+
+          drawInfoArr.push(convertChange);
+        }
+      } else {
+        let markerImg = '';
+
+        if (properties.pointType === 'S') {
+          // 출발지 마커
+          markerImg = 'http://tmapapi.sktelecom.com/upload/tmap/marker/pin_r_m_s.png';
+        } else if (properties.pointType === 'E') {
+          // 도착지 마커
+          markerImg = 'http://tmapapi.sktelecom.com/upload/tmap/marker/pin_r_m_e.png';
+        } else {
+          // 각 포인트 마커
+          markerImg = `http://tmapapi.sktelecom.com/upload/tmap/marker/pin_b_m_${viaPoint}.png`;
+          viaPoint += 1;
+        }
+
+        // 경로들의 결과값들을 포인트 객체로 변환
+        const latlon = new Tmapv3.Point(geometry.coordinates[0], geometry.coordinates[1]);
+        // 포인트 객체를 받아 좌표값으로 다시 변환
+        const convertPoint = new Tmapv3.Projection.convertEPSG3857ToWGS84GEO(latlon);
+
+        resultMarkerArr.push({
+          position: {
+            lat: Number(convertPoint._lat),
+            lng: Number(convertPoint._lng),
+          },
+          content: properties.viaPointName,
+          image: markerImg,
+        });
+        bounds.extend(new kakao.maps.LatLng(Number(convertPoint._lat), Number(convertPoint._lng)));
+      }
+    }
+    // 검색된 장소 위치를 기준으로 지도 범위를 재설정합니다
+    map?.setBounds(bounds);
+    setPath(drawInfoArr);
+    setPreviewMarkers(resultMarkerArr);
+  };
 
   const addLocation = (marker: MarkerProps) => {
     selected.push(marker);
     setSelected(selected);
   };
 
-  // useEffect(() => {
-  //   console.log('reload');
-  // }, [courseState.tMapData]);
-  //
-  // const fetchRouteData = async () => {
-  //   // await dispatch(fetchPathFromTMap());
-  //   // setResultData(courseState.tMapData);
-  //   // setResultFeatures(courseState.tMapFeatures);
-  //   // console.log(courseState.tMapData);
-  //   processData();
-  //   setPreview(false);
-  // };
+  useEffect(() => {
+    if (!map) return;
+    console.log(preview);
+    if (preview) {
+      dispatch(fetchPathFromTMap());
+      setResultData(courseState.tMapData);
+      setResultFeatures(courseState.tMapFeatures);
+      // setResultData(poisData.properties);
+      // setResultFeatures(poisData.features);
+    }
+  }, [preview, map, courseState]);
+
+  useEffect(() => {
+    if (resultData && resultFeatures) {
+      processData();
+    }
+  }, [resultFeatures, resultData]);
 
   return (
     <div className="Container" style={{ display: 'flex', position: 'fixed' }}>
@@ -206,7 +212,7 @@ export default function CourseCreate() {
         </button>
       </div>
       <SearchBar
-        markers={markers}
+        markers={searchMarkers}
         selected={selected}
         searchPlaces={searchPlaces}
         setInfo={setInfo}
@@ -217,9 +223,9 @@ export default function CourseCreate() {
       <KakaoMap
         map={map}
         setMap={setMap}
-        // path={path}
-        markers={markers}
-        setMarkers={setMarkers}
+        path={path}
+        searchMarkers={searchMarkers}
+        previewMarkers={previewMarkers}
         info={info}
         setInfo={setInfo}
         addLocation={addLocation}
